--- src/legacies/assetManagement/SpendingGoal/components/MonthSpendingGoal/components/ModifyModal/index.ts
@@ -1,3 +1,0 @@
-import ModifyModal from "@legacies/assetManagement/SpendingGoal/components/MonthSpendingGoal/components/ModifyModal/ModifyModal.tsx";
-
-export default ModifyModal;
