--- src/legacies/assetManagement/SpendingGoal/components/MonthSpendingGoal/index.ts
@@ -1,3 +1,0 @@
-import MonthSpendingGoal from "@legacies/assetManagement/SpendingGoal/components/MonthSpendingGoal/MonthSpendingGoal.tsx";
-
-export default MonthSpendingGoal;
