--- src/legacies/assetManagement/SpendingGoal/components/RegularSpendingGoal/components/ConfirmModal/index.ts
@@ -1,3 +1,0 @@
-import ConfirmModal from "@legacies/assetManagement/SpendingGoal/components/RegularSpendingGoal/components/ConfirmModal/ConfirmModal.tsx";
-
-export default ConfirmModal;
