--- src/legacies/assetManagement/SpendingGoal/components/RegularSpendingGoal/components/GoalCard/index.ts
@@ -1,3 +1,0 @@
-import GoalCard from "@legacies/assetManagement/SpendingGoal/components/RegularSpendingGoal/components/GoalCard/GoalCard.tsx";
-
-export default GoalCard;
