--- src/legacies/assetManagement/SpendingGoal/components/RegularSpendingGoal/index.ts
@@ -1,3 +1,0 @@
-import RegularSpendingGoal from "@legacies/assetManagement/SpendingGoal/components/RegularSpendingGoal/RegularSpendingGoal.tsx";
-
-export default RegularSpendingGoal;
