--- src/legacies/assetManagement/SpendingGoal/index.ts
@@ -1,3 +1,0 @@
-import SpendingGoal from "@legacies/assetManagement/SpendingGoal/SpendingGoal.tsx";
-
-export default SpendingGoal;
