+++ src/pages/AssetManagement/pages/SpendingGoal/components/MonthSpendingGoal/components/ModifyModal/index.ts
@@ -1,0 +1,3 @@
+import ModifyModal from "@pages/AssetManagement/pages/SpendingGoal/components/MonthSpendingGoal/components/ModifyModal/ModifyModal.tsx";
+
+export default ModifyModal;
