+++ src/pages/AssetManagement/pages/SpendingGoal/components/MonthSpendingGoal/index.ts
@@ -1,0 +1,3 @@
+import MonthSpendingGoal from "@pages/AssetManagement/pages/SpendingGoal/components/MonthSpendingGoal/MonthSpendingGoal.tsx";
+
+export default MonthSpendingGoal;
