+++ src/pages/AssetManagement/pages/SpendingGoal/components/RegularSpendingGoal/components/ConfirmModal/index.ts
@@ -1,0 +1,3 @@
+import ConfirmModal from "@pages/AssetManagement/pages/SpendingGoal/components/RegularSpendingGoal/components/ConfirmModal/ConfirmModal.tsx";
+
+export default ConfirmModal;
