+++ src/pages/AssetManagement/pages/SpendingGoal/components/RegularSpendingGoal/components/GoalCard/index.ts
@@ -1,0 +1,3 @@
+import GoalCard from "@pages/AssetManagement/pages/SpendingGoal/components/RegularSpendingGoal/components/GoalCard/GoalCard.tsx";
+
+export default GoalCard;
