+++ src/pages/AssetManagement/pages/SpendingGoal/components/RegularSpendingGoal/index.ts
@@ -1,0 +1,3 @@
+import RegularSpendingGoal from "@pages/AssetManagement/pages/SpendingGoal/components/RegularSpendingGoal/RegularSpendingGoal.tsx";
+
+export default RegularSpendingGoal;
