+++ src/pages/AssetManagement/pages/SpendingGoal/index.ts
@@ -1,0 +1,3 @@
+import SpendingGoal from "@pages/AssetManagement/pages/SpendingGoal/SpendingGoal.tsx";
+
+export default SpendingGoal;
